Move TF-related constants into panels/ThreeDimensionalViz
